+++ ckan/config/routing.py
@@ -9,10 +9,10 @@ refer to the routes manual at http://routes.groovie.org/docs/
 """
 import re
 
-from pylons import config
 from routes.mapper import SubMapper, Mapper as _Mapper
 
 import ckan.plugins as p
+from ckan.common import config
 
 named_routes = {}
 
@@ -427,6 +427,9 @@ def make_map():
         m.connect('/util/redirect', action='redirect')
         m.connect('/testing/primer', action='primer')
         m.connect('/testing/markup', action='markup')
+
+    # robots.txt
+    map.connect('/(robots.txt)', controller='template', action='view')
 
     # Mark all unmarked routes added up until now as core routes
     for route in map.matchlist: