+++ ckan/logic/action/delete.py
@@ -2,8 +2,11 @@
 
 '''API functions for deleting data from CKAN.'''
 
+import logging
+
 import sqlalchemy as sqla
 
+import ckan.lib.jobs as jobs
 import ckan.logic
 import ckan.logic.action
 import ckan.plugins as plugins
@@ -11,6 +14,9 @@ import ckan.lib.dictization.model_dictize as model_dictize
 from ckan import authz
 
 from ckan.common import _
+
+
+log = logging.getLogger('ckan.logic')
 
 validate = ckan.lib.navl.dictization_functions.validate
 
@@ -50,7 +56,7 @@ def user_delete(context, data_dict):
     user.delete()
 
     user_memberships = model.Session.query(model.Member).filter(
-        model.Member.table_id == user_id).all()
+        model.Member.table_id == user.id).all()
 
     for membership in user_memberships:
         membership.delete()
@@ -701,3 +707,48 @@ def unfollow_group(context, data_dict):
             ckan.logic.schema.default_follow_group_schema())
     _unfollow(context, data_dict, schema,
             context['model'].UserFollowingGroup)
+
+
+@ckan.logic.validate(ckan.logic.schema.job_clear_schema)
+def job_clear(context, data_dict):
+    '''Clear background job queues.
+
+    Does not affect jobs that are already being processed.
+
+    :param list queues: The queues to clear. If not given then ALL
+        queues are cleared.
+
+    :returns: The cleared queues.
+    :rtype: list
+
+    .. versionadded:: 2.7
+    '''
+    _check_access(u'job_clear', context, data_dict)
+    queues = data_dict.get(u'queues')
+    if queues:
+        queues = [jobs.get_queue(q) for q in queues]
+    else:
+        queues = jobs.get_all_queues()
+    names = [jobs.remove_queue_name_prefix(queue.name) for queue in queues]
+    for queue, name in zip(queues, names):
+        queue.empty()
+        log.info(u'Cleared background job queue "{}"'.format(name))
+    return names
+
+
+def job_cancel(context, data_dict):
+    '''Cancel a queued background job.
+
+    Removes the job from the queue and deletes it.
+
+    :param string id: The ID of the background job.
+
+    .. versionadded:: 2.7
+    '''
+    _check_access(u'job_cancel', context, data_dict)
+    id = _get_or_bust(data_dict, u'id')
+    try:
+        jobs.job_from_id(id).delete()
+        log.info(u'Cancelled background job {}'.format(id))
+    except KeyError:
+        raise NotFound