+++ ckan/tests/test_coding_standards.py
@@ -219,8 +219,10 @@ def find_unprefixed_string_literals(filename):
                         break
                     except ValueError:
                         continue
-            first_char = lines[lineno][col_offset]
-            if first_char not in u'ub':  # Don't allow capital U and B either
+            leading = lines[lineno][col_offset - 1:col_offset + 1]
+            if leading[:-1] == u'[':  # data['id'] is unambiguous, ignore these
+                continue
+            if leading[-1:] not in u'ub':  # Don't allow capital U and B either
                 problems.append((lineno + 1, col_offset + 1))
     return sorted(problems)
 
@@ -252,7 +254,6 @@ _STRING_LITERALS_WHITELIST = [
     u'ckan/controllers/revision.py',
     u'ckan/controllers/storage.py',
     u'ckan/controllers/tag.py',
-    u'ckan/controllers/template.py',
     u'ckan/controllers/user.py',
     u'ckan/controllers/util.py',
     u'ckan/exceptions.py',
@@ -426,7 +427,6 @@ _STRING_LITERALS_WHITELIST = [
     u'ckan/model/vocabulary.py',
     u'ckan/pastertemplates/__init__.py',
     u'ckan/plugins/core.py',
-    u'ckan/plugins/interfaces.py',
     u'ckan/plugins/toolkit.py',
     u'ckan/plugins/toolkit_sphinx_extension.py',
     u'ckan/tests/config/test_environment.py',